build: Extract jacoco custom plugin's functions
- Create extension functions for most configure blocks defined within
  the custom jacoco plugin.

Resolves: DCMAW-7147
