+++ buildLogic/plugins/src/main/kotlin/uk.gov.ui.jacoco-module-config.gradle.kts
@@ -1,15 +1,10 @@
-import com.android.build.api.variant.LibraryAndroidComponentsExtension
 import com.android.build.gradle.LibraryExtension
-import uk.gov.ui.ext.BaseVariantExt.capitalisedFlavorName
-import uk.gov.ui.filetree.fetcher.AsmFileTreeFetcher
-import uk.gov.ui.filetree.fetcher.FileTreesFetcher
-import uk.gov.ui.filetree.fetcher.JavaCompileFileTreeFetcher
-import uk.gov.ui.filetree.fetcher.KotlinCompileFileTreeFetcher
-import uk.gov.ui.jacoco.tasks.JacocoCombinedTestTaskGenerator
-import uk.gov.ui.jacoco.tasks.JacocoConnectedTestTaskGenerator
-import uk.gov.ui.jacoco.tasks.JacocoTaskGenerator
-import uk.gov.ui.jacoco.tasks.JacocoUnitTestTaskGenerator
+import uk.gov.ui.ext.decorateDslLibraryExtensionWithJacoco
+import uk.gov.ui.ext.decorateLibraryExtensionWithJacoco
+import uk.gov.ui.ext.decorateTestTasksWithJacoco
+import uk.gov.ui.ext.generateDebugJacocoTasks
 import com.android.build.api.dsl.LibraryExtension as DslLibraryExtension
+import uk.gov.ui.ext.debugLog
 
 plugins {
     jacoco
@@ -18,100 +13,34 @@ plugins {
 val jacocoVersion: String by rootProject.extra
 
 project.tasks.withType<Test> {
-    if (this.hasProperty("jacoco")) {
-        (this.property("jacoco") as JacocoTaskExtension).let {
-            it.isIncludeNoLocationClasses = true
-            it.excludes = listOf("jdk.internal.*")
-        }
+    decorateTestTasksWithJacoco().also {
+        project.debugLog("Applied jacoco properties to Test tasks")
     }
 }
 
 project.configure<JacocoPluginExtension> {
     this.toolVersion = jacocoVersion
+    project.debugLog("Applied jacoco tool version to jacoco plugin")
 }
 
 project.configure<LibraryExtension> {
-    testCoverage.jacocoVersion = jacocoVersion
-    buildTypes {
-        debug {
-            this.enableAndroidTestCoverage = true
-            this.enableUnitTestCoverage = true
-            this.isTestCoverageEnabled = true
-        }
+    decorateLibraryExtensionWithJacoco(jacocoVersion).also {
+        project.debugLog("Applied jacoco properties to Library")
     }
 }
 
 project.configure<DslLibraryExtension> {
-    testCoverage.jacocoVersion = jacocoVersion
-    buildTypes {
-        debug {
-            this.enableAndroidTestCoverage = true
-            this.enableUnitTestCoverage = true
-            this.isTestCoverageEnabled = true
-        }
-    }
-}
-
-project.configure<LibraryAndroidComponentsExtension> {
-    this.onVariants { debugLibraryVariant ->
-        println("MC: Library variant: ${debugLibraryVariant.name}")
+    decorateDslLibraryExtensionWithJacoco(jacocoVersion).also {
+        project.debugLog("Applied jacoco properties to DSL Library")
     }
 }
 
 project.afterEvaluate {
     (this.findProperty("android") as LibraryExtension).let { extension ->
-
-        extension.libraryVariants.forEach { debugLibraryVariant ->
-            if (debugLibraryVariant.buildType.name == "debug") {
-                val variantName = debugLibraryVariant.name
-
-                val classDirectoriesFetcher = FileTreesFetcher(
-                    project,
-                    KotlinCompileFileTreeFetcher(
-                        project,
-                        variantName,
-                        debugLibraryVariant.capitalisedFlavorName,
-                    ),
-                    AsmFileTreeFetcher(
-                        project,
-                        variantName,
-                        debugLibraryVariant.capitalisedFlavorName,
-                    ),
-                    JavaCompileFileTreeFetcher(
-                        project,
-                        variantName,
-                        debugLibraryVariant.capitalisedFlavorName,
-                    ),
-                )
-
-                val unitTestReportGenerator = JacocoUnitTestTaskGenerator(
-                    project,
-                    classDirectoriesFetcher,
-                    variantName,
-                )
-
-                val connectedTestReportGenerator = JacocoConnectedTestTaskGenerator(
-                    project,
-                    classDirectoriesFetcher,
-                    variantName,
-                )
-
-                val combinedTestReportGenerator = JacocoCombinedTestTaskGenerator(
-                    project = project,
-                    classDirectoriesFetcher = classDirectoriesFetcher,
-                    variantName = variantName,
-                    configs = listOf(
-                        unitTestReportGenerator,
-                        connectedTestReportGenerator,
-                    ),
-                )
-
-                listOf(
-                    unitTestReportGenerator,
-                    connectedTestReportGenerator,
-                    combinedTestReportGenerator,
-                ).forEach(JacocoTaskGenerator::customTask)
-            }
+        extension.libraryVariants.forEach {
+            it.generateDebugJacocoTasks(this)
         }
+
+        project.debugLog("Applied jacoco custom tasks")
     }
 }
