+++ buildLogic/plugins/src/main/kotlin/uk/gov/ui/ext/ProjectExt.kt
@@ -1,0 +1,7 @@
+package uk.gov.ui.ext
+
+import org.gradle.api.Project
+
+fun Project.debugLog(messageSuffix: String) {
+    logger.debug("${project.path}: $messageSuffix")
+}
